requirements.txt
pandas 
pyabsa
seqeval
supabase
beautifulsoup4
transformers
googletrans
google-play-scraper
python-dotenv
